Update App, Event, and New Event Form components
