--- cypress.json
@@ -1,10 +1,0 @@
-{
-  "cypress-watch-and-reload": {
-    "watch": "src/**/*.js"
-  },
-  "env": {
-    "cypress-react-selector": {
-      "root": ".__cy_root"
-    }
-  }
-}
