+++ package.json
@@ -3,14 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
-    "uuid": "^8.3.2",
-    "web-vitals": "^2.1.4"
+    "react-scripts": "4.0.3",
+    "web-vitals": "^1.1.2",
+    "uuid": "^9.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -37,8 +37,8 @@
     ]
   },
   "devDependencies": {
-    "@cypress/react": "^6.1.1",
-    "cypress": "^9.7.0",
+    "@cypress/react": "^6.2.0",
+    "cypress": "^10.10.0",
     "cypress-react-selector": "^3.0.0",
     "cypress-watch-and-reload": "^1.10.3"
   }
