+++ src/App.js
@@ -5,66 +5,13 @@ import { v1 as generateUniqueID } from "uuid";
 // import Event from "./Components/Event";
 import Footer from "./Components/Footer";
 import Header from "./Components/Header";
+import NewEventForm from "./Components/NewEventForm";
 // import NewEventForm from "./Components/NewEventForm";
 
 function App() {
   const [events, setEvents] = useState(eventsData);
 
   const [showAttendees, setShowAttendees] = useState(false);
-
-  const [selectOption, setSelectOption] = useState("");
-
-  const [newEvent, setNewEvent] = useState({
-    id: "",
-    eventType: "",
-    name: "",
-    organizer: "",
-    eventImage: "",
-    date: "",
-    people: [],
-  });
-
-  function addEvent() {
-    const createEvent = {
-      id: generateUniqueID(),
-      eventType: selectOption,
-      name: newEvent.name,
-      organizer: newEvent.organizer,
-      eventImage: newEvent.eventImage || "https://loremflickr.com/640/480/",
-      date: newEvent.date,
-      people: [],
-    };
-    handleAddEvent(createEvent);
-  }
-
-  function handleSelectChange(e) {
-    setSelectOption(e.target.value);
-  }
-
-  function handleSubmit(e) {
-    e.preventDefault();
-    addEvent();
-    resetEventForm();
-  }
-
-  function handleTextChange(e) {
-    setNewEvent({
-      ...newEvent,
-      [e.target.id]: e.target.value,
-    });
-  }
-
-  function resetEventForm() {
-    setNewEvent({
-      id: "",
-      eventType: "",
-      name: "",
-      organizer: "",
-      eventImage: "",
-      date: "",
-    });
-    setSelectOption("");
-  }
 
   function handleAddEvent(event) {
     setEvents([event, ...events]);
@@ -93,45 +40,7 @@ function App() {
       <Header />
       <main>
         <div className="new-event">
-          <>
-            <form onSubmit={handleSubmit}>
-              <h3>Create a new event</h3>
-              <label htmlFor="name">Event name:</label>
-              <input
-                type="text"
-                id="name"
-                onChange={handleTextChange}
-                value={newEvent.name}
-              />
-
-              <label htmlFor="organizer">Organizer:</label>
-              <input
-                type="text"
-                id="organizer"
-                onChange={handleTextChange}
-                value={newEvent.organizer}
-              />
-
-              <label htmlFor="eventImage">Event image:</label>
-              <input
-                type="text"
-                id="eventImage"
-                onChange={handleTextChange}
-                value={newEvent.eventImage}
-              />
-              <label htmlFor="eventType">Event type:</label>
-              <select id="eventType" onChange={handleSelectChange}>
-                <option value=""></option>
-                <option value="Birthday">Birthday</option>
-                <option value="Anniversary">Anniversary</option>
-                <option value="Intramural Sport">Intramural Sport</option>
-                <option value="Watch Party">Watch Party</option>
-                <option value="wedding">Wedding</option>
-              </select>
-              <br />
-              <input type="submit" />
-            </form>
-          </>
+          <NewEventForm />
         </div>
         <div className="events">
           <ul>
@@ -140,6 +49,7 @@ function App() {
 
               return (
                 <>
+                  {/* Event */}
                   <li key={event.id}>
                     <img src={event.eventImage} alt={event.name} />
                     <h5>
@@ -149,6 +59,7 @@ function App() {
                     <span>Organized by: {event.organizer} </span>
                     <br />
                     <>
+                      {/* Attendees */}
                       <button onClick={toggleEventAttendees}>
                         {!showAttendees ? "Show Attendees" : "Hide Attendees"}
                       </button>
@@ -157,6 +68,7 @@ function App() {
                         <div className="attendees">
                           {attendees.map((attendee, index) => (
                             <>
+                              {/* Attendee */}
                               <div key={attendee.id} className="attendee">
                                 <p>
                                   <img
